pyproject.toml
[project]
name = "notifyservice"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi[standard]>=0.111.0",
  "uvicorn[standard]>=0.35.0",
  "redis>=6.4.0",
  "PyJWT>=2.10.1",
  "pydantic-settings>=2.8.1"
]

[tool.uvicorn]
factory = true
app = "app.main:create_app"
host = "0.0.0.0"
port = 8000